Fix Kitsu `synopsis` nullability (#1233)
This time, the Kitsu API docs are silent on whether this field (or
any other field) can be null/undefined/etc, but it can happen and
caused an error during search and update. This change just ensures the
attribute is nullable and is set to an empty String when it is null.

(cherry picked from commit 96f1605e5e2286aa8cc4c7ed4d203db61d7c631b)